+++ test_job.go
@@ -1,5 +1,5 @@
 package main
 
 type TestJob struct {
-	costPredictionSeconds int
+	costPredictionSeconds   float64
 }
